+++ vite-project/src/App.css
@@ -250,6 +250,7 @@ body {
   display: flex;
   align-items: flex-start;
   justify-content: space-between;
+  margin-bottom: 22px;
 }
 .pop-browse__form {
   max-width: 370px;
@@ -278,6 +279,7 @@ body {
 .form-browse__block {
   display: flex;
   flex-direction: column;
+  margin-right: 20px;
 }
 .form-browse__area {
   max-width: 370px;
